+++ solid/frame/file/tempbase.hpp
@@ -19,12 +19,12 @@ namespace file {
 
 struct TempBase {
     const size_t   tempstorageid;
-    const size_t   tempid;
+    const uint32_t tempid;
     const uint64_t tempsize;
 
     TempBase(
         size_t   _storageid,
-        size_t   _id,
+        uint32_t _id,
         uint64_t _size)
         : tempstorageid(_storageid)
         , tempid(_id)
@@ -43,8 +43,8 @@ struct TempBase {
 
     virtual void close(const char* _path, bool _remove) = 0;
 
-    virtual int read(char* _pb, uint32_t _bl, int64_t _off)        = 0;
-    virtual int write(const char* _pb, uint32_t _bl, int64_t _off) = 0;
+    virtual ssize_t read(char* _pb, size_t _bl, int64_t _off)        = 0;
+    virtual ssize_t write(const char* _pb, size_t _bl, int64_t _off) = 0;
     virtual int64_t size() const = 0;
     virtual bool truncate(int64_t _len = 0) = 0;
     virtual void flush();
