+++ solid/system/seekabledevice.hpp
@@ -20,9 +20,9 @@ public:
     using Device::read;
     using Device::write;
     //! Read from a given position
-    int read(char* _pb, size_t _bl, int64_t _off);
+    ssize_t read(char* _pb, size_t _bl, int64_t _off);
     //! Write at a given position
-    int write(const char* _pb, size_t _bl, int64_t _off);
+    ssize_t write(const char* _pb, size_t _bl, int64_t _off);
     //! Move the cursor to a given position
     int64_t seek(int64_t _pos, SeekRef _ref = SeekBeg);
     //! Truncate to a certain length
